test: replace jest with uvu
