+++ package.json
@@ -9,7 +9,7 @@
   "types": "./index.d.ts",
   "scripts": {
     "test": "npm run -s test:code && npm run -s test:typescript",
-    "test:code": "jest --coverage",
+    "test:code": "node test.js",
     "test:typescript": "npx tsc --noEmit --declaration --noUnusedLocals test-typescript.ts"
   },
   "repository": "github:gr2m/javascript-plugin-architecture-with-typescript-definitions",
@@ -27,22 +27,8 @@
   ],
   "license": "ISC",
   "devDependencies": {
-    "@types/jest": "^26.0.23",
-    "@types/node": "^14.0.1",
-    "jest": "^27.0.4",
-    "ts-jest": "^27.0.3",
-    "typescript": "^4.0.0"
-  },
-  "jest": {
-    "preset": "ts-jest",
-    "coverageThreshold": {
-      "global": {
-        "statements": 100,
-        "branches": 100,
-        "functions": 100,
-        "lines": 100
-      }
-    }
+    "typescript": "^4.0.0",
+    "uvu": "^0.5.1"
   },
   "renovate": {
     "extends": [
